【PAC】feat：开启PAC模式的代码库支持自动同步代码库YAML变更到蓝盾 #8130
merge后删除版本分支

+++ src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/yaml/PipelineYamlRepositoryService.kt
@@ -321,12 +321,14 @@ class PipelineYamlRepositoryService @Autowired constructor(
                     branchVersionAction = BranchVersionAction.INACTIVE
                 )
             }
-            pipelineYamlService.deleteBranchFile(
-                projectId = projectId,
-                repoHashId = repoHashId,
-                branch = branch,
-                filePath = filePath
-            )
+            if (!yamlFile.ref.isNullOrBlank()) {
+                pipelineYamlService.deleteBranchFile(
+                    projectId = projectId,
+                    repoHashId = repoHashId,
+                    branch = yamlFile.ref,
+                    filePath = filePath
+                )
+            }
         } catch (ignored: Exception) {
             logger.error("Failed to delete pipeline yaml|$projectId|${action.format()}", ignored)
             throw ignored
